+++ benches/note_decryption.rs
@@ -3,7 +3,7 @@ use orchard::{
     builder::Builder,
     bundle::Flags,
     circuit::ProvingKey,
-    keys::{FullViewingKey, Scope, SpendingKey},
+    keys::{FullViewingKey, PreparedIncomingViewingKey, Scope, SpendingKey},
     note_encryption::{CompactAction, OrchardDomain},
     value::NoteValue,
     Anchor, Bundle,
@@ -21,6 +21,7 @@ fn bench_note_decryption(c: &mut Criterion) {
     let fvk = FullViewingKey::from(&SpendingKey::from_bytes([7; 32]).unwrap());
     let valid_ivk = fvk.to_ivk(Scope::External);
     let recipient = valid_ivk.address_at(0u32);
+    let valid_ivk = PreparedIncomingViewingKey::new(&valid_ivk);
 
     // Compact actions don't have the full AEAD ciphertext, so ZIP 307 trial-decryption
     // relies on an invalid ivk resulting in random noise for which the note commitment
@@ -31,15 +32,15 @@ fn bench_note_decryption(c: &mut Criterion) {
     //
     // Our fixed (action, invalid ivk) tuple will always fall into a specific rejection
     // case. In order to reflect the real behaviour in the benchmarks, we trial-decrypt
-    // with 1000 invalid ivks (each of which will result in a different uniformly-random
-    // plaintext); this is equivalent to trial-decrypting 1000 different actions with the
+    // with 10240 invalid ivks (each of which will result in a different uniformly-random
+    // plaintext); this is equivalent to trial-decrypting 10240 different actions with the
     // same ivk, but is faster to set up.
-    let invalid_ivks: Vec<_> = (0u32..1000)
+    let invalid_ivks: Vec<_> = (0u32..10240)
         .map(|i| {
             let mut sk = [0; 32];
             sk[..4].copy_from_slice(&i.to_le_bytes());
             let fvk = FullViewingKey::from(&SpendingKey::from_bytes(sk).unwrap());
-            fvk.to_ivk(Scope::External)
+            PreparedIncomingViewingKey::new(&fvk.to_ivk(Scope::External))
         })
         .collect();